#Angular-QuizeApp
I developed this project as part of the Self-interest.

Quiz-App is an Angular app in which users can play quiz nearly 10 multiple choice questions in total. This repository is made for the purpose to showcase my project for learning purposes and should not be used commercially.

The salient features of the app are:
   1. Login is provided to the user so that only that person can access the app.
   2. Total Timer for the quiz .
   3. User can see his score at the end of exam. 
   
I also highly appreciate if you could suggest some improvements in the project.